refactor(filters): Split out math filters
